==> doc/RMD/09.Rmd <==
---
title: "R Notebook"
output:
  html_document:
    df_print: paged
    toc: yes
  html_notebook:
    toc: yes
editor_options:
  chunk_output_type: inline
---

# WMN Differentiation

```{r setup, include = FALSE}
PROJHOME <- rprojroot::find_rstudio_root_file()
knitr::opts_knit$set(root.dir = PROJHOME)
```



```{r packages, message = FALSE, warning = FALSE}
library("tidyverse")
library("poppr")
(load("data/sclerotinia_16_loci.rda"))
```

# Purpose

The white mold nursery populations are unique because they are not fungicide
treated and have the same cultivars planted in them year after year.

The question becomes, are white mold nurseries differentiated from each other or
are they more or less homogeneous? We could use AMOVA to test for these with
location and binary source (wmn or non-wmn) as the hierarchy.


# Data Setup

First, we want to clone-correct our data down to the field level so that we
don't accidentally include non-independent samples.

```{r correction}
dat11
dat11cc <- clonecorrect(dat11, ~Region/Source/Host/Year)
dat11cc

make_amova_table <- function(am, amt, samples = "Region"){
  tot <- nrow(am$results)
  res <- data.frame(list(am$results[-tot, c("Df", "Sum Sq")], 
                         Percent = am$componentsofcovariance[-tot, 2],
                         Pval    = rev(amt$pvalue), 
                         # Sigma   = rev(amt$)
                         Phi     = rev(am$statphi$Phi[-tot])))
  res <- as.matrix(res)
  colnames(res) <- c("d.f.", "Sum of Squares", "Percent variation", "P", 
                     "Phi statistic")
  names(dimnames(res)) <- c("levels", "statistic")
  rownames(res) <- gsub("samples", samples, rownames(res))
  return(res)
}

make_amova_printable <- function(amtab, amtabcc){
  am_array <- array(dim      = c(dim(amtab), 2),
                    dimnames = c(dimnames(amtab), 
                                 list(c("full", "clone-corrected"))))
  am_array[, , 1] <- amtab
  am_array[, , 2] <- amtabcc
  tabfun <- function(x){
    x <- paste0(paste0(signif(x, 3), collapse = " ("), ")")
    return(x)
  }
  res <- apply(am_array, c(1, 2), tabfun)
  return(res)
}
```

Now that we've done that, we should make a new variable in the strata that 
separates the white mold nurseries from the others. We'll call this stratum 
"Source Type".

```{r new_stratum}
addStrata(dat11cc) <- strata(dat11cc) %>% 
  mutate(SourceType = forcats::fct_inorder(ifelse(Source == "wmn", "wmn", "other"))) %>%
  select(SourceType)
setPop(dat11cc) <- ~SourceType
dat11cc
```

I can perform AMOVA on the newly defined variable using Bruvo's distance.

```{r AMOVA, cache = TRUE}
other(dat11cc)$REPLEN
bd <- bruvo.dist(dat11cc, replen = other(dat11cc)$REPLEN)
(ssc_amova <- poppr.amova(dat11cc, ~SourceType, dist = bd, quiet = TRUE))
ssc_amova_test <- randtest(ssc_amova, nrepet = 999)
plot(ssc_amova_test)
ssc_amova_test
```


This result is telling us that there is some subdivision between white mold
nurseries and non-white mold nurseries. Of course, from previous analyses, we
know that Mexico is differentiated from other populations, so what happens if we
account for Region? Here, we are placing region lower in the heirarchy because
we specifically want to test the effect of region on the differentiation between
white mold nurseries within different regions.

```{r AMOVA-Region, cache = TRUE}
ssc_amova_region <- poppr.amova(dat11cc, ~SourceType/Region, dist = bd, quiet = TRUE)
ssc_amova_region_test <- randtest(ssc_amova_region, nrepet = 9999)
plot(ssc_amova_region_test)
(ssc_amova_table <- make_amova_table(ssc_amova_region, ssc_amova_region_test))
```

Okay! This shows that when we account for Region after accounting for Source
Type, we find that the differentiation is coming mainly from the Regions. What
happens when we remove Mexico?

```{r AMOVA-nomex, cache = TRUE}
datnomex <- setPop(dat11cc, ~Region) %>% popsub(blacklist = "Mexico")
bdnm     <- bruvo.dist(datnomex, replen = other(datnomex)$REPLEN)
ssc_amova_nm <- poppr.amova(datnomex, ~SourceType/Region, dist = bdnm, quiet = TRUE)
ssc_amova_nm_test <- randtest(ssc_amova_nm, nrepet = 9999)
plot(ssc_amova_nm_test)
(ssc_amova_nm_table <- make_amova_table(ssc_amova_nm, ssc_amova_nm_test))
```

When we remove the Mexican isolates (which only contained white mold nurseries
and shared no genotypes), we see that indeed, the degree of differentiation
went down.  Of course, if we look at the distribution of isolates between white
mold nurseries and regions, we can see things are a bit lopsided:

```{r wmntable}
table(strata(dat11cc, ~SourceType/Region, combine = FALSE))
```

The only regions that have at least ten samples in both white mold nurseries and
production regions are NE, WA, and MI. We can see what happens when we subsample
to these


```{r AMOVA-newami, cache = TRUE}
datnewami <- setPop(dat11cc, ~Region) %>% popsub(sublist = c("NE", "WA", "MI"))
bdnewami     <- bruvo.dist(datnewami, replen = other(datnewami)$REPLEN)
ssc_amova_newami <- poppr.amova(datnewami, ~SourceType/Region, dist = bdnewami, quiet = TRUE)
ssc_amova_newami_test <- randtest(ssc_amova_newami, nrepet = 9999)
plot(ssc_amova_newami_test)
(ssc_amova_newami_table <- make_amova_table(ssc_amova_newami, ssc_amova_newami_test))
```

```{r AMOVA_table, comment = NA}
make_amova_printable(ssc_amova_table, ssc_amova_newami_table) %>%  
  as_tibble() %>%
  add_column(Hierarchy = c("Between Source", "Between Region within Source", "Within Region"), .before = 1) %>%
  readr::write_csv(path = file.path(PROJHOME, "results", "tables", "AMOVA-region.csv"), col_names = TRUE) %>%
  rename(ps = `Phi statistic`) %>%
  mutate(ps = gsub("0\\.00(\\d{1})(\\d{2})", "\\1.\\2e^-3^", ps)) %>%
  mutate(ps = case_when(Hierarchy == "Between Source" ~ ps, TRUE ~ paste0("**", ps, "**"))) %>%
  rename(`$\\Phi statistic$` = ps) %>%
  rename(`% variation` = `Percent variation`) %>%
  rename(S.S. = `Sum of Squares`) %>%
  select(-P) %>%
  huxtable::as_huxtable(add_colnames = TRUE) %>% 
  huxtable::set_col_width(c(1.1, 0.6, 0.8, 0.8, 1.1)) %>% 
  huxtable::set_align(huxtable::everywhere, 2:5, "center") %>% 
  huxtable::print_md(max_width = 90) 
```



## Visualizing the partitions

We can visualize the partitions if we create distributions showing the genetic
distance.

```{r distrib, fig.width = 6, fig.height = 3}
wmn_inds <- (setPop(dat11cc, ~SourceType) %>% pop()) == "wmn"

# Function to set the upper triangle to NA to avoid over-representation.
set_upper_tri_NA <- function(d){d[upper.tri(d)] <- NA; diag(d) <- NA; d}

# dist data converted to matrix fed into here. Removes NAs from previous function
tidy_dist <- . %>%
  as.data.frame() %>% 
  rownames_to_column("from") %>% 
  gather(to, distance, -from) %>% 
  filter(!is.na(distance))
wmn_distance   <- as.matrix(bd)[wmn_inds, wmn_inds] %>% set_upper_tri_NA() %>% tidy_dist
nwmn_distance  <- as.matrix(bd)[!wmn_inds, !wmn_inds] %>% set_upper_tri_NA() %>% tidy_dist
inter_distance <- as.matrix(bd)[wmn_inds, !wmn_inds] %>% tidy_dist

dists <- bind_rows(`White Mold Nurseries` = wmn_distance, 
                   `Between` = inter_distance,
                   `Other Sources` = nwmn_distance, 
                   .id = "Comparison") %>%
  mutate(Comparison = forcats::fct_inorder(Comparison)) %>%
  as_tibble()
dists
ggplot(dists, aes(y = distance, x = Comparison, fill = Comparison)) +
  geom_violin() +
  geom_boxplot(width = 0.25) +
  scale_fill_manual(values =  c(`Other Sources` = "grey35", `White Mold Nurseries` = "grey95", Between = "grey65")) +
  scale_color_manual(values = c(`Other Sources` = "grey35", `White Mold Nurseries` = "grey95", Between = "grey65")) +
  theme_bw(base_size = 16, base_family = "Helvetica") +
  theme(aspect.ratio = 1/2) +
  theme(legend.position = "none") +
  theme(axis.title.x = element_blank()) +
  theme(axis.text = element_text(color = "black")) +
  theme(panel.grid.major = element_line(colour = "grey20")) +
  theme(panel.grid.minor = element_line(colour = "grey50")) +
  theme(panel.grid.major.x = element_blank()) +
  theme(panel.border = element_blank()) +
  ylab("Bruvo's Distance")
```

```{r dist_comparison, results = "asis"}
summary_dists <- dists %>% 
  group_by(Comparison) %>% 
  summarize(`Average Distance` = mean(distance),
            `Standard Deviation` = sd(distance)) 
knitr::kable(summary_dists, digits = 3)
```

On first look, it appears that there's not much difference. Looking at the
table, we can see that the average distance within White Mold Nurseries is 
`r round(summary_dists[["Average Distance"]][1], 3)`, whereas the average distance within Other
Sources is `r round(summary_dists[["Average Distance"]][3], 3)`. This is well within the standard
deviation for both distributions, but we also recognize that a single step of
Bruvo's distance for 11 loci in a haploid organism is `r round(0.5/11, 3)`.


## WMN DAPC

DAPC is a nice way of visualizing these. First, I'm simply going to compare 
between white mold nurseries and others:

```{r wmndapc, cache = TRUE}
filter <- strata(dat11cc) %>%
  group_by(Region) %>%
  mutate(filter = length(unique(SourceType)) > 1) %>%
  pull(filter)
wmn_compare <- dat11cc[filter]
set.seed(2017-07-12)
wmndapc <- xvalDapc(tab(wmn_compare), pop(wmn_compare), n.rep = 200, n.pca = 1:10)
wmndapc$DAPC
scatter(wmndapc$DAPC)
(wmntab <- table(wmndapc$DAPC$assign, wmndapc$DAPC$grp))
chisq.test(wmntab, simulate.p.value = TRUE)
```

Okay, this shows that there's quite a bit of overlap between these two. What
happens if I look at the DAPC when it includes the pouplation groupings. Mainly,
I want to know if the white mold nurseries can be compared while considering the
pouplations

```{r wmndapcpop, fig.width = 8, fig.height = 10, cache = TRUE}
set.seed(2017-07-12)
wmndapc <- xvalDapc(tab(wmn_compare), strata(wmn_compare, ~SourceType/Region)[[2]], 
                    n.pca = 5:20, n.rep = 100)$DAPC
scatter(wmndapc)
LDS <- bind_cols(Population = wmndapc$grp, as.data.frame(wmndapc$ind.coord)) %>%
  as_tibble()
LDS_pop <- LDS %>% 
  group_by(Population) %>% 
  summarize_all(mean) %>%
  rename_all(function(x) gsub("LD", "mean", x))
LDS <- full_join(LDS, LDS_pop) %>%
  separate(Population, c("SourceType", "Pop"), remove = FALSE) 
LDS_PLOT <- ggplot(LDS, aes(x = LD1, y = LD2, color = SourceType)) + 
  geom_point(aes(fill = SourceType), alpha = 0.5, pch = 21, color = "black") +
  geom_segment(aes(x = mean1, y = mean2, xend = LD1, yend = LD2), alpha = 0.5) +
  stat_ellipse(type = "norm", level = 0.66, alpha = 0.75) + 
  # ggrepel::geom_label_repel(aes(x = mean1, y = mean2, label = Population), 
  #                           data = LDS_pop, show.legend = FALSE, color = "black") +
  theme_bw() +  
  theme(aspect.ratio = 1/1.618) +
  theme(legend.position = "bottom") +
  theme(axis.text = element_blank()) + 
  theme(axis.title = element_blank()) + 
  theme(axis.ticks = element_blank()) + 
  viridis::scale_color_viridis(discrete = TRUE, direction = -1) +
  viridis::scale_fill_viridis(discrete = TRUE, direction = -1) +
  scale_y_continuous(breaks = 0) + 
  scale_x_continuous(breaks = 0) + 
  theme(panel.background = element_rect(fill = "grey95")) +
  theme(panel.grid.major = element_line(color = "black"))

LDS_PLOT + facet_wrap(~Pop, ncol = 3)

```



# Host differentiation

There are three main hosts shared within the white mold nurseries: 

| Cultivar | Susceptibility |
| ----- | ---- |
| Beryl | Susceptible |
| Bunsi | ? |
| G122  | Some Resistance |

These hosts are grown in the nurseries at different times and it is of interest
if cultivar can have an effect on population structure. Given the results of
Aldrich-Wolfe 2015, I would suspect that cultivar has no effect, but it's
important to verify this. We are going to use AMOVA and DAPC for this.

```{r host_AMOVA}
bebug <- dat11cc %>% 
  setPop(~Source) %>% 
  popsub("wmn") %>% 
  setPop(~Host) %>% 
  popsub(c("Beryl", "Bunsi", "G122"))
bebugd <- bruvo.dist(bebug, replen = other(bebug)$REPLEN)
(ssc_host_amova <- poppr.amova(bebug, ~Host, dist = bebugd, quiet = TRUE))
(ssc_host_amova <- poppr.amova(bebug, ~Host/Region, dist = bebugd, quiet = TRUE))
```

The AMOVA makes it clear that there's no differentiation between hosts. Let's
see what DAPC shows us.

```{r dapc, fig.width = 5, fig.height = 4.5}
setPop(bebug) <- ~Host
bebug_dapc <- dapc(bebug, n.pca = 20, n.da = 2)
scatter(bebug_dapc, #clabel = 0,
        legend = TRUE,
        scree.pca = TRUE, 
        scree.da = TRUE,
        col = viridis::plasma(3, end = 0.85),
        bg.inset = "grey90",
        bg = "grey90")

ggcompoplot::ggcompoplot(bebug_dapc, setPop(bebug, ~Region), cols = 5, pal = viridis::plasma(3, end = 0.85)) +
  theme(legend.position = "top")
apply(bebug_dapc$posterior, 2, mean)
```

```{r dapc_hr, fig.width = 5, fig.height = 5.5}
setPop(bebug) <- ~Host/Region
bebug_dapc <- dapc(bebug, n.pca = 20, n.da = 20)
bebug_dapc
```

```{r ggdapc, fig.width = 8, fig.height = 4}
LDS <- bind_cols(Population = bebug_dapc$grp, as.data.frame(bebug_dapc$ind.coord)) %>%
  as_tibble()
LDS_pop <- LDS %>% 
  group_by(Population) %>% 
  summarize_all(mean) %>%
  rename_all(function(x) gsub("LD", "mean", x))
LDS <- full_join(LDS, LDS_pop, by = "Population") %>% separate(Population, c("Cultivar", "Population"))
LDS_PLOT <- ggplot(LDS, aes(x = LD1, y = LD2, color = Cultivar)) + 
  geom_point(aes(fill = Cultivar), alpha = 0.5, pch = 21, color = "black") +
  geom_segment(aes(x = mean1, y = mean2, xend = LD1, yend = LD2), alpha = 0.5) +
  stat_ellipse(type = "norm", level = 0.66, alpha = 0.75) + 
  theme_bw(base_size = 16, base_family = "Helvetica") +  
  theme(aspect.ratio = 1/1.618) +
  theme(legend.position = "bottom") +
  theme(axis.text = element_blank()) + 
  theme(axis.title = element_blank()) + 
  theme(axis.ticks = element_blank()) + 
  viridis::scale_color_viridis(option = "B", discrete = TRUE, direction = -1, end = 0.85) +
  viridis::scale_fill_viridis(option = "B", discrete = TRUE, direction = -1, end = 0.85) +
  scale_y_continuous(breaks = 0) + 
  scale_x_continuous(breaks = 0) + 
  theme(panel.background = element_rect(fill = "grey95")) +
  theme(panel.grid.major = element_line(color = "black")) +
  facet_wrap(~Population, nrow = 2)
LDS_PLOT
```



<details>
<summary>Session Information</summary>

```{r, echo = FALSE}
options(width = 100)
devtools::session_info()
```

</details>
